+++ drivers/net/ethernet/intel/idpf/Makefile
@@ -19,4 +19,5 @@ idpf-y := \
 	idpf_vf_dev.o		\
 	idpf_virtchnl.o 	\
 	idpf_vport.o		\
-	idpf_vport_config.o
+	idpf_vport_config.o	\
+	idpf_vport_q.o
